Attempted to update for filling in tracks missed in big projects
Didn't quite work I don't think however.  But, REAPER's console doesn't seem to print anything so I can't quite locate the issue.  If anyone wants, they can have a go at trying to find out where the problem.  We're trying to, in bigger projects where the mouse is too fast for the tick rate, make sure the tracks along the way are counted as well, and that none are skipped.